feat(map,connections): add a basic view to display connection information + jumps

+++ db/src/main/scala/org/updraft0/controltower/db/model/map.scala
@@ -102,7 +102,7 @@ case class MapWormholeConnectionJump(
     connectionId: ConnectionId,
     characterId: CharacterId,
     shipTypeId: Int,
-    massOverride: Option[Int],
+    massOverride: Option[Long],
     createdAt: Instant
 )
