+++ protocol/shared/src/main/scala/org/updraft0/controltower/protocol/map.scala
@@ -80,7 +80,7 @@ case class MapWormholeConnectionJump(
     connectionId: ConnectionId,
     characterId: CharacterId,
     shipTypeId: Int,
-    massOverride: Option[Int],
+    massOverride: Option[Long],
     createdAt: Instant
 )
